# Add GitHub Pages deployment workflow and performance fixes