--- a/ai_seo_kit_v1/sample-frontmatter.md
+++ b/ai_seo_kit_v1/sample-frontmatter.md
@@ -1,0 +1,13 @@
+---
+title: "Sample Post Title"
+slug: sample-post-title
+description: "A concise description of the post for SEO and social sharing."
+tags: ["seo", "ai", "freelancing"]
+primary_keyword: "ai seo for freelancers"
+intent: "informational"
+schema_type: "BlogPosting"
+canonical: "/blog/sample-post-title"
+date: "2025-01-01"
+---
+
+This is an example body of a blog post. Replace this text with your actual content. Make sure to include the primary keyword in the first sentence and maintain readability.