--- a/test_agent.txt
+++ b/test_agent.txt
@@ -1,0 +1,1 @@
+Test file created by agent client