--- a/test_agent_output.txt
+++ b/test_agent_output.txt
@@ -1,0 +1,3 @@
+Test file created by agent client
+Line 2
+Line 3