--- a/test_diff.txt
+++ b/test_diff.txt
@@ -1,0 +1,1 @@
+Test file created by agent client