… .github/workflows/gh-pages.yml
… ai_seo_kit_v1/30-day-playbook.md
… ai_seo_kit_v1/json-ld-snippets.json
… ai_seo_kit_v1/templates/audit-template.md
… ai_seo_kit_v1/templates/guest-post-pitch-template.md
… ai_seo_kit_v1/templates/linkedin-post-template.md
… ai_seo_kit_v1/templates/outreach-email-template.md
… ai_seo_kit_v1/templates/review-response-template.md
… ai_seo_kit_v1/templates/twitter-thread-template.md
… perf-budget.json
… scripts/README_AGENT.md
… scripts/__pycache__/agent_client.cpython-311.pyc
… scripts/__pycache__/bridge.cpython-311.pyc
… scripts/setup_agent.sh
… scripts/test_agent.py
… seo-portfolio/content/ai-seo-kit/30-day-playbook.md
… seo-portfolio/content/ai-seo-kit/json-ld-snippets.json
… seo-portfolio/content/ai-seo-kit/sample-frontmatter.md
… seo-portfolio/content/ai-seo-kit/templates/audit-template.md
… seo-portfolio/content/ai-seo-kit/templates/guest-post-pitch-template.md
… seo-portfolio/content/ai-seo-kit/templates/linkedin-post-template.md
… seo-portfolio/content/ai-seo-kit/templates/outreach-email-template.md
… seo-portfolio/content/ai-seo-kit/templates/review-response-template.md
… seo-portfolio/content/ai-seo-kit/templates/twitter-thread-template.md
… seo-portfolio/perf-budget.json
… seo-portfolio/public/downloads/ai-seo-kit-v1.zip
… seo-portfolio/scripts/review-response-generator.js
… seo-portfolio/src/content/blog/30-day-ai-seo-playbook.md
… seo-portfolio/src/pages/products/ai-seo-kit.md
… seo-portfolio/src/pages/products/buy-ai-seo-kit.md